Fix an issue with array access.
The case of accessing an array with a list was not implemented.  For
example, `@array[0, 1, 2]`.

Unfortunately, there are probably other ways that arrays and hashes can
be access that are also not properly implemented yet.  So there will
probably need to be more work done for this yet.

Also version bump so that this can be published.